refactor: profileのコンポーネントを他機能のコンポーネントのファイル構造に統一する、それに伴う参照ファイル側でのimportパス修正 #208
