+++ frontend/src/features/profile/layout/profileLayout.tsx
@@ -3,9 +3,9 @@ import { RiCompass3Line } from "react-icons/ri";
 import styles from "@/features/profile/layout/profileLayout.module.css";
 import { Modal, type ModalHandle } from "@/components/ui/modal/modal";
 import { useRef } from "react";
-import { ProfileCard } from "../components/profileCard";
-import { OtherProfileCard } from "../components/otherProfileCard";
-import { ProfileCardTab } from "../components/profileTab";
+import { ProfileCard } from "../components/profileCard/profileCard";
+import { OtherProfileCard } from "../components/profileCard/otherProfileCard";
+import { ProfileCardTab } from "../components/profileTab/profileTab";
 import { Outlet } from "react-router";
 import { NoticeInfo } from "@/features/searchClassroom/components/noticeInfo";
 import { type NoticeInfoItem } from "@/features/searchClassroom/types/SelectClassroom";
